sum of multiples of 3 and 5 up to but excluding 1024

+++ 0x02-functions_nested_loops/101-natural.c
@@ -8,7 +8,7 @@
 
 int main(void)
 {
-	long i, sum = 0;
+	int i, sum = 0;
 
 	for (i = 0; i < 1024; i++)
 	{
